# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ClearCut API
  version: 1.0.0
  description: Screenplay Pre-Clearance Evidence Workspace API
servers:
  - url: /
    description: Current origin
paths:
...
  /api/v1/organizations/{orgId}/projects/{projectId}:
    get:
      summary: Get details of a project
      operationId: getProject
      tags:
        - projects
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/organizations/{orgId}/projects/{projectId}/clearance-items/{itemId}:research:
    post:
      summary: Start automated Parallel search & extract research for a clearance item
      operationId: startResearch
      tags:
        - research
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
        - name: itemId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
      responses:
        '202':
          description: Research job started
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Job'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
...
  /api/v1/organizations/{orgId}/projects/{projectId}/clearance-items/{itemId}:assign:
    post:
      summary: Assign clearance item to an authorized project member
      operationId: assignClearanceItem
      tags:
        - review
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
        - name: itemId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignClearanceItemRequest'
      responses:
        '200':
          description: Item assigned with the resulting item version
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ClearanceItem'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '403':
          description: Actor lacks the current assignment capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Item or assignee is not found in the authorized project scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale item version or idempotency-key intent mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Assignment command validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/organizations/{orgId}/projects/{projectId}/clearance-items/{itemId}:changeDueDate:
    post:
      summary: Change due date of clearance item
      operationId: changeClearanceItemDueDate
      tags:
        - review
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
        - name: itemId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - dueDate
              properties:
                dueDate:
                  $ref: '#/components/schemas/ISODateTime'
      responses:
        '200':
          description: Due date changed
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ClearanceItem'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
  /api/v1/organizations/{orgId}/projects/{projectId}/clearance-items/{itemId}:recordEvidenceDecision:
    post:
      summary: Record a human reviewer's treatment of cited evidence
      description: Records an evidence review outcome for pre-clearance workflow only; it is not a legal-clearance conclusion.
      operationId: recordEvidenceDecision
      tags:
        - decisions
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
        - name: itemId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordEvidenceDecisionRequest'
      responses:
        '200':
          description: Evidence decision recorded with the resulting item version
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ClearanceItem'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '403':
          description: Actor lacks the current evidence-decision capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Item is not found in the authorized project scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale item version or idempotency-key intent mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Decision, rationale, intent, or evidence policy validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/organizations/{orgId}/projects/{projectId}/clearance-items/{itemId}:setDisposition:
    post:
      summary: Set the human pre-clearance workflow disposition for an item
      description: Records a governed workflow disposition without guaranteeing legal clearance.
      operationId: setDisposition
      tags:
        - decisions
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
        - name: itemId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDispositionRequest'
      responses:
        '200':
          description: Disposition recorded with the resulting item version
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/ClearanceItem'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '403':
          description: Actor lacks the current disposition capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Item is not found in the authorized project scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale item version or idempotency-key intent mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Disposition, rationale, intent, or evidence policy validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/organizations/{orgId}/projects/{projectId}/clearance-items/{itemId}:refer:
    post:
      summary: Submit an item referral to an authorized specialist role
      operationId: referClearanceItem
      tags:
        - collaboration
      security:
        - SessionCookie: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ProjectId'
        - name: itemId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferClearanceItemRequest'
      responses:
        '201':
          description: Referral submitted with the resulting item version
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    $ref: '#/components/schemas/Referral'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '403':
          description: Actor lacks the current referral capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Item or referral target is not found in the authorized project scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stale item version or idempotency-key intent mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Referral command validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      description: Organization UUIDv7
      schema:
        $ref: '#/components/schemas/UUIDv7'
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Project UUIDv7
      schema:
        $ref: '#/components/schemas/UUIDv7'
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: clearcut_session
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    UUIDv7:
      type: string
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      description: UUID version 7 identifier
...
    Project:
      type: object
      required:
        - projectId
        - orgId
        - title
        - productionType
        - productionStage
        - jurisdiction
        - targetLockDate
        - reviewBrief
        - createdAt
      properties:
        projectId:
          $ref: '#/components/schemas/UUIDv7'
        orgId:
          $ref: '#/components/schemas/UUIDv7'
        title:
          type: string
        description:
          anyOf:
            - type: string
            - type: 'null'
        productionType:
          anyOf:
            - type: string
            - type: 'null'
        productionStage:
          anyOf:
            - type: string
            - type: 'null'
        jurisdiction:
          anyOf:
            - type: string
            - type: 'null'
        targetLockDate:
          anyOf:
            - type: string
              format: date
            - type: 'null'
        reviewBrief:
          anyOf:
            - type: string
            - type: 'null'
        isDemo:
          type: boolean
          default: false
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
...
    ResponseMeta:
      type: object
      required:
        - requestId
      properties:
        requestId:
          type: string
        nextCursor:
          type: string
          nullable: true
        totalCount:
          type: integer
          nullable: true
...